Modularized SGD and added run script.
I added a script that runs the SGD procedure.  This, or a similar
script is what you should now use to run experiments.

The “runnable” functionality has been removed from sgd.py.  You should
use the functions from this module when you run the procedure.  Only
bugfixes to this module from now on!